requirement.txt
speech_recognition
pyaudio
gtts
mysql.connector
google.generativeai
webbrowser
smtplib
pywhatkit
playsound
keyboard
spotipy
time
datetime
os
